% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_information.R
\name{account_information}
\alias{account_information}
\title{Function enables you to get information regarding your Hunter account at any time. This call is free.}
\usage{
account_information(key)
}
\arguments{
\item{key}{One of the API keys associated with your account.
account_information()}
}
\description{
This function allows you to get information regarding your Hunter account at any time. This call is free.
}
