% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_search.R
\name{domain_search}
\alias{domain_search}
\title{Function to get every email address found on the internet using a given domain name, with sources.}
\usage{
domain_search(
  domain,
  key,
  limit = "",
  offset = "",
  type = "",
  seniority = "",
  department = ""
)
}
\arguments{
\item{domain}{The domain you want to get the mails for}

\item{key}{Your hunter.io-API key}

\item{limit}{Specifies the max number of email addresses to return. The default is 10.}

\item{offset}{Specifies the number of email addresses to skip. The default is 0.}

\item{type}{Get only personal or generic email addresses.}

\item{seniority}{Get only email addresses for people with the selected seniority level. The possible values are junior, senior or executive. Several seniority levels can be selected (delimited by a comma).}

\item{department}{Get only email addresses for people working in the selected department(s). The possible values are executive, it, finance, management, sales, legal, support, hr, marketing or communication. Several departments can be selected (comma-delimited).
domain_search()}
}
\description{
This function allows you to get the email adress found on the web for a given domain you enter.
}
