% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_count.R
\name{email_count}
\alias{email_count}
\title{Function to get how many email addresses hunter.io have for one domain or for one company.}
\usage{
email_count(domain, type = "")
}
\arguments{
\item{type}{Get the count of only personal or generic email addresses.
email_count()}

\item{email}{The email address you want to verify.}
}
\description{
This function allows you to know how many email addresses we have for one domain or for one company. It's free and doesn't require authentication.
}
