% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_verifier.R
\name{email_verifier}
\alias{email_verifier}
\title{Function to verify the deliverability of an email address.}
\usage{
email_verifier(email, key)
}
\arguments{
\item{email}{The email address you want to verify.}

\item{key}{Your hunter.io-API key
email_verifier()}
}
\description{
This function allows you to verify the deliverability of an email address.
}
\examples{
\dontrun{
email_verifier("danielatkeyword-hero.com", key)
}
}
